+++ pkgs/dart_mcp/lib/src/client/elicitation_support.dart
@@ -4,19 +4,27 @@
 
 part of 'client.dart';
 
-/// The interface for handling elicitation requests.
-///
-/// Any client using [ElicitationSupport] must implement this interface.
+@Deprecated(
+  'This interface is going away, the method will exist directly on the '
+  'ElicitationSupport mixin instead',
+)
 abstract interface class WithElicitationHandler {
   FutureOr<ElicitResult> handleElicitation(ElicitRequest request);
 }
 
 /// A mixin that adds support for the `elicitation` capability to an
 /// [MCPClient].
+// ignore: deprecated_member_use_from_same_package
 base mixin ElicitationSupport on MCPClient implements WithElicitationHandler {
   @override
   void initialize() {
     capabilities.elicitation ??= ElicitationCapability();
     super.initialize();
   }
+
+  /// The method for handling elicitation requests.
+  ///
+  /// Any client using [ElicitationSupport] must implement this interface.
+  @override
+  FutureOr<ElicitResult> handleElicitation(ElicitRequest request);
 }